Describe API types as YAML.
There doesn't appear to be a schema definition for the API.  This is a rough
capture of the type information from the docs.

It could probably be an OpenAPI spec, but the level of detail that requires is
beyond my patience right now. Eventually I would like to generate the types
from the schema, but I haven't gotten that far yet.